refactor: Phase 2 — split AdminComponents.tsx into one file per component
AdminComponents.tsx (966 lines, 8 components/helpers in one file) is now
src/components/admin/AdminComponents/{UsersTable,CompaniesTable,
DefaultsForm,SubmissionsTable}.tsx + _shared.tsx for ActionBtn/FormField/
Section2/style constants. index.ts re-exports the four named exports so
all 4 existing call sites (admin/{users,companies,defaults,submissions}/
page.tsx) import from @/components/admin/AdminComponents unchanged.

While splitting, swapped the local type duplicates for the Phase 1 shared
types per audit item [Architecture] AdminComponents.tsx:82-90,201-211,743-756:
- CompanyRow -> ATSCompanyRow (lib/types.ts already had the same shape
  plus created_at/updated_at; CompanyForm is now
  Omit<ATSCompanyRow, "id" | "created_at" | "updated_at">)
- SubmissionRow -> ATSSubmission (lib/types.ts version is a strict
  superset - adds city/reviewed_at/reviewed_by, narrows ats_type and
  status to their real union types)
- UserRow -> AdminUserListItem, with one correction: Phase 1 typed
  user_settings as a single object, but the actual Supabase embed
  (and the component's existing u.user_settings?.[0]?.uses_defaults)
  treats it as an array. Fixed the type in lib/types.ts to match
  reality instead of changing working component logic.

Also deduped ATS_TYPES, a byte-for-byte copy of lib/constants.ts'
VALID_ATS that the Phase 1 pass missed because it only touched API
route files, not this component.

No styling or accessibility changes - that's Phase 3, done in the same
pass as the htmlFor/focus-trap fixes per the fix plan.

Validation: tsc --noEmit clean, 65/65 tests passing, next build
successful (all 4 admin routes compile via the new index.ts re-export),
lint-staged (eslint --fix + prettier) clean with zero further edits.
