+++ src/components/admin/AdminComponents/index.ts
@@ -1,0 +1,8 @@
+// src/components/admin/AdminComponents/index.ts
+// Re-exports so existing `@/components/admin/AdminComponents` imports keep
+// working unchanged after the file-per-component split.
+
+export { UsersTable } from "./UsersTable";
+export { CompaniesTable } from "./CompaniesTable";
+export { DefaultsForm } from "./DefaultsForm";
+export { SubmissionsTable } from "./SubmissionsTable";
